requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.0.11
Django==1.10
django-allauth==0.32.0
django-appconf==1.0.2
django-autocomplete-light==3.2.2
django-braces==1.11.0
django-classy-tags==0.8.0
django-easy-maps==0.9.3
django-generic-positions==0.2.2
django-geoip==0.5.2.1
django-hvad==1.7.0
django-ipgeobase==1.0.2b0
django-libs==1.67.12
django-location-field==2.0.3
django-model-utils==3.0.0
django-ratings==0.3.7
django-star-ratings==0.5.5
django-tinymce==2.6.0
django-user-media==1.2.4
easy-thumbnails==2.4.1
geocoder==1.22.4
geoip2==2.4.2
geopy==1.11.0
gunicorn==19.7.1
ipaddress==1.0.18
maxminddb==1.3.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
ratelim==0.1.6
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
Unidecode==0.4.20
